Benchmark CRS generation with two parties (#368)
This is a simple two-party benchmark of CRS generation. Running on one machine vs. more than one machine should not make too much difference since the computation is sequential, i.e., one party generates the CRS and then sends it to another party to verify.

Using n = 57249 as specified in page 68 of https://eprint.iacr.org/2023/800.pdf, the CRS size works out to be just below 12 MB after deserialization using bincode.

The time for two rounds (which is the minimum) is 115 seconds on a M3 Macbook Pro on a single thread. This is a big chunk of time (20 minutes for 10 samples) if we plan to run this on the CI, so maybe it's a good idea to set n to a lower number since the computation time scales linearly.

This PR also updates the CRS ceremony code to create the proof in a separate thread using `rayon::spawn` so that it does not block the async executioner.