feat(gateway): simplify reenc POST API (#736)
* feat(gateway): simplify reenc POST API

* fix(gateway): use existing ciphertext if exists

* feat: Refactor blockchain module structure
Refactor the blockchain module structure to improve code organization and maintainability. Renamed the `decrypt` module to `handlers` and updated the import statements accordingly. This change also required updating the references to the `handle_event_decryption` and `handle_reencryption_event` functions in the `events/manager.rs` file.

* chore: Add todo!()s for Euint512 and Euint1024 in kms.proto, rpc_types.rs, and kms.rs

Update the `kms.proto`, `rpc_types.rs`, and `kms.rs` files to add support for the `Euint512` and `Euint1024` FheTypes. This includes adding the enum variants, updating the message structures, and implementing the necessary conversions and handling logic in the codebase.

* chore: Remove unused code in ciphertext_provider.rs and handlers.rs

Remove unused code in the `ciphertext_provider.rs` and `handlers.rs` files. This includes removing the unused `FheType::from(ct_handle[30])` line in `ciphertext_provider.rs` and the unused loop variable in `handlers.rs`.

* feat(gateway): add chain ID to config

* chore(wasm): integrate insecure reconstruction into wasm

* chore(gateway): better logging

---------

Co-authored-by: Jay Logelin <jay.logelin@gmail.com>