feat: Update variable name in handle_reencryption_event (#745)
* feat: Update variable name in handle_reencryption_event

The variable name `ciphertext_digest` has been updated to `ciphertext_handle` in the `handle_reencryption_event` function in the `handlers.rs` file. This change aligns with the corresponding change in the `ApiReencryptValues` struct in the `manager.rs` file.

* chore: Update variable name in handle_reencryption_event

The variable name `ciphertext_digest` has been updated to `ciphertext_handle` in the `handle_reencryption_event` function in the `handlers.rs` file. This change aligns with the corresponding change in the `ApiReencryptValues` struct in the `manager.rs` file.