feat(gateway): Refactor reenc api service (#774)
* feat(gateway): Add chain ID to reencryption request

This commit adds the `chain_id` field to the reencryption request in the `Blockchain` trait and its implementation. It also updates the `gateway.toml` configuration file to use a new chain ID value.

* feat: Add chain ID to reencryption request and update gateway configuration

The code changes add the `chain_id` field to the reencryption request in the `Blockchain` trait and its implementation. It also updates the `gateway.toml` configuration file to use a new chain ID value.

* chore: make clippy happy, formatting