feat(core): crs ceremony for v2 proofs (#2088)
V2 CRS ceremony now takes ~50 seconds, this is much faster than before which took ~10-15 minutes locally. This is because the witness dimension is 10x smaller. From the CRS ceremony perspective, the difference is only the witness dimension. Which version to use is chosen based on the compact public key parameter.

Closes #2041