chore: upgrade tfhe-rs to v1 (#2095)
Note that this is not a complete upgrade as the securely generated keys are not guaranteed to work (maybe the computation will be wrong, maybe they'll be right, we don't know). This is because we don't support the generation of the mod switch key yet, and it will be done in another issue #2094.

Centrally generate and insecurely generated keys should come with the new mod switch key.

Closes #2087